Make Pytorch->Flax conversion use specified dtype

+++ src/diffusers/models/modeling_flax_utils.py
@@ -416,7 +416,7 @@ class FlaxModelMixin(PushToHubMixin):
             pytorch_model_file = load_state_dict(model_file)
 
             # Step 2: Convert the weights
-            state = convert_pytorch_state_dict_to_flax(pytorch_model_file, model)
+            state = convert_pytorch_state_dict_to_flax(pytorch_model_file, model, dtype=dtype)
         else:
             try:
                 with open(model_file, "rb") as state_f:
